% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidrology.R
\name{elev_to_hidrology}
\alias{elev_to_hidrology}
\title{Calculate derivatives from the Hidrology module}
\usage{
elev_to_hidrology(
  elev_sgrd = NULL,
  slope_sgrd = NULL,
  spcar_sgrd = NULL,
  out_dir,
  prefix = "",
  env = env,
  ...,
  mridx = FALSE,
  sagaw = FALSE,
  sllgt = FALSE,
  spidx = FALSE
)
}
\arguments{
\item{elev_sgrd}{optional input, elevation raster data in SAGA format,
can be created with \code{elev_to_sgrd()}}

\item{slope_sgrd}{optional input, slope raster data in SAGA format,
can be created with \code{elev_to_morphometry(slope = T)}}

\item{spcar_sgrd}{optional input, specific catchment area raster data in
SAGA format, can be created with
\code{elev_to_terrain_analysis(flow = TRUE, spcar = T)}
Both options are needed since this is computed from the
Flow Accumulation (One Step) module}

\item{out_dir}{output directory}

\item{prefix}{character prefix for output filenames}

\item{env}{environment to get SAGA installation,
can be set with \code{init_saga()}}

\item{...}{ignored, check help page for possible outputs}

\item{mridx}{boolean, defaults to \code{FALSE}}

\item{sagaw}{boolean, defaults to \code{FALSE}}

\item{sllgt}{boolean, defaults to \code{FALSE}}

\item{spidx}{boolean, defaults to \code{FALSE}}
}
\description{
Calculate derivatives from the Hidrology module
}
