% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting.R
\name{elev_to_lighting}
\alias{elev_to_lighting}
\title{Calculate derivatives from the Lighting module}
\usage{
elev_to_lighting(
  elev_sgrd,
  out_dir,
  prefix = "",
  env = env,
  ...,
  negop = FALSE,
  posop = FALSE,
  shade = FALSE,
  svfct = FALSE,
  visky = FALSE
)
}
\arguments{
\item{elev_sgrd}{input, elevation raster data in SAGA format,
can be created with \code{elev_to_sgrd()}}

\item{out_dir}{output directory}

\item{prefix}{character prefix for output filenames}

\item{env}{environment to get SAGA installation,
can be set with \code{init_saga()}}

\item{negop}{boolean, defaults to \code{FALSE}}

\item{posop}{boolean, defaults to \code{FALSE}}

\item{shade}{boolean, defaults to \code{FALSE}}

\item{svfct}{boolean, defaults to \code{FALSE}}

\item{visky}{boolean, defaults to \code{FALSE}}
}
\description{
Calculate derivatives from the Lighting module
}
