% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphometry.R
\name{elev_to_morphometry}
\alias{elev_to_morphometry}
\title{Calculate derivatives from the Morphometry module}
\usage{
elev_to_morphometry(
  elev_sgrd,
  out_dir,
  prefix = "",
  env = env,
  ...,
  aspect = FALSE,
  ccros = FALSE,
  cgene = FALSE,
  clong = FALSE,
  cmaxi = FALSE,
  cmini = FALSE,
  cplan = FALSE,
  cprof = FALSE,
  cvidx = FALSE,
  ddgrd = FALSE,
  mbidx = FALSE,
  mdslp = FALSE,
  nrght = FALSE,
  slhgt = FALSE,
  slope = FALSE,
  sthgt = FALSE,
  textu = FALSE,
  tpidx = FALSE,
  tridx = FALSE,
  vldpt = FALSE,
  units = "degrees"
)
}
\arguments{
\item{elev_sgrd}{input, elevation raster data in SAGA format,
can be created with \code{elev_to_sgrd()}}

\item{out_dir}{output directory}

\item{prefix}{character prefix for output filenames}

\item{env}{environment to get SAGA installation,
can be set with \code{init_saga()}}

\item{...}{ignored, check help page for possible outputs}

\item{aspect}{boolean, defaults to \code{FALSE}}

\item{ccros}{boolean, defaults to \code{FALSE}}

\item{cgene}{boolean, defaults to \code{FALSE}}

\item{clong}{boolean, defaults to \code{FALSE}}

\item{cmaxi}{boolean, defaults to \code{FALSE}}

\item{cmini}{boolean, defaults to \code{FALSE}}

\item{cplan}{boolean, defaults to \code{FALSE}}

\item{cprof}{boolean, defaults to \code{FALSE}}

\item{cvidx}{boolean, defaults to \code{FALSE}}

\item{ddgrd}{boolean, defaults to \code{FALSE}}

\item{mbidx}{boolean, defaults to \code{FALSE}}

\item{mdslp}{boolean, defaults to \code{FALSE}}

\item{slhgt}{boolean, defaults to \code{FALSE}}

\item{slope}{boolean, defaults to \code{FALSE}}

\item{sthgt}{boolean, defaults to \code{FALSE}}

\item{textu}{boolean, defaults to \code{FALSE}}

\item{tpidx}{boolean, defaults to \code{FALSE}}

\item{tridx}{boolean, defaults to \code{FALSE}}

\item{vldpt}{boolean, defaults to \code{FALSE}}

\item{units}{character, units for slope and aspect output}

\item{nrhgt}{boolean, defaults to \code{FALSE}}
}
\description{
Calculate derivatives from the Morphometry module
}
