% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_analysis.R
\name{elev_to_terrain_analysis}
\alias{elev_to_terrain_analysis}
\title{Calculate derivatives from the Terrain Analysis module}
\usage{
elev_to_terrain_analysis(
  elev_sgrd,
  out_dir,
  prefix = "",
  env = env,
  ...,
  lsfct = FALSE,
  flow = FALSE,
  spcar = FALSE,
  twidx = FALSE
)
}
\arguments{
\item{elev_sgrd}{input, elevation raster data in SAGA format,
can be created with \code{elev_to_sgrd()}}

\item{out_dir}{output directory}

\item{prefix}{character prefix for output filenames}

\item{env}{environment to get SAGA installation,
can be set with \code{init_saga()}}

\item{...}{ignored, check help page for possible outputs}

\item{lsfct}{boolean, defaults to \code{FALSE}}

\item{flow}{boolean, defaults to \code{FALSE}}

\item{spcar}{boolean, defaults to \code{FALSE} When \code{TRUE},
then \code{flow} should also be \code{TRUE}}

\item{twidx}{boolean, defaults to \code{FALSE}}
}
\description{
Calculate derivatives from the Terrain Analysis module
}
