% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_saga}
\alias{init_saga}
\title{Initiate SAGA environment}
\usage{
init_saga(path, modules = NULL)
}
\arguments{
\item{path}{path to saga portable installation}

\item{modules}{path to saga modules, defaults to the tools directory in the SAGA path}
}
\description{
Initiates the SAGA environment before running any other RSAGA function.
Will set an "env" object to the R Global Environment.
}
