% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{terrain_to_tif}
\alias{terrain_to_tif}
\title{Batch convert resulting derivatives to GeoTIFF}
\usage{
terrain_to_tif(out_dir_to_translate, out_crs = NULL)
}
\arguments{
\item{out_dir_to_translate}{directory containing outputs from
\code{elev_to_*()}. Resulting TIFF files will be
saved in this directory as well.}

\item{out_crs}{CRS information is lost within SAGA processes, it is therefore
the need to reset it at the end of the process. Here one can
give CRS information of the original elevation file used to
create the derivatives. This can be obtained with
\code{terra::crs(elev_filename)}. If the original elevation
file does not have a CRS, then pass the corresponding CRS
information as WKT, PROJ.4 or EPSG values.}
}
\description{
Assuming several derivatives have been computed, this function will map over
directory of resulting terrain derivatives and translate the files into TIFF
format if not yet existing. To translate a single file from SAGA to GeoTIFF
format better use \code{gdalUtils::gdal_translate()}.
}
\details{
If the elevation file resulting from \code{elev_to_sgrd()} is
present in the out_dir_to_translate, it will give a warning and will
not convert the file to TIF, since the driver is corrupted.
To refer to elevation data always use the original file.
}
