% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{elev_to_channel}
\alias{elev_to_channel}
\title{Calculate derivatives from the Channel module}
\usage{
elev_to_channel(
  elev_sgrd,
  flow_sgrd,
  out_dir,
  prefix = "",
  env = env,
  ...,
  chnet = TRUE,
  vdcnw = FALSE
)
}
\arguments{
\item{elev_sgrd}{input, elevation raster data in SAGA format,
can be created with \code{elev_to_sgrd()}}

\item{flow_sgrd}{input, can be generated with \code{elev_to_terrain_analysis()}}

\item{out_dir}{output directory}

\item{prefix}{character prefix for output filenames}

\item{env}{environment to get SAGA installation,
can be set with \code{init_saga()}}

\item{chnet}{boolean, defaults to \code{FALSE}}

\item{vdcnw}{boolean, defaults to \code{FALSE}}
}
\description{
The main goal of this function is to calculate the
Vertical Distance to the Channel Network \code{vdcnw}
parameter, so it will calculate first \code{chnet}.
However, it can be used to only calcualte the
Channel Network as well.
}
