% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{elev_to_channel}
\alias{elev_to_channel}
\title{Calculate derivatives from the Channel module}
\usage{
elev_to_channel(
  elev_sgrd,
  flow_sgrd,
  out_dir,
  prefix = "",
  envir,
  ...,
  chnet = TRUE,
  vdcnw = FALSE,
  init_value = 1e+06,
  chnet_shp = F
)
}
\arguments{
\item{elev_sgrd}{input, elevation raster data in SAGA format,
can be created with \code{elev_to_sgrd()}}

\item{flow_sgrd}{input, can be generated with
\code{elev_to_terrain_analysis()}}

\item{out_dir}{output directory}

\item{prefix}{character prefix for output filenames}

\item{envir}{environment to get SAGA installation,
can be set with \code{init_saga()}}

\item{...}{ignored, check help page for possible outputs}

\item{chnet}{Channel Network:
If a cell is part of a channel its value equals the channel
order. Otherwise the cell is marked as no-data.
Boolean, defaults to \code{TRUE}}

\item{vdcnw}{Vertical Distance to Channel Network:
Altitude above the channel network.
Boolean, defaults to \code{FALSE}}

\item{init_value}{Initiation flow accumulation value for
channel network computation. Defaults to 1000000}

\item{chnet_shp}{Calculate channel network in vector format}
}
\description{
The main goal of this function is to calculate the
Vertical Distance to the Channel Network \code{vdcnw}
parameter, so it will calculate first \code{chnet}.
However, it can be used to only calculate the
Channel Network as well.
}
