% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidrology.R
\name{elev_to_hidrology}
\alias{elev_to_hidrology}
\title{Calculate derivatives from the Hidrology module}
\usage{
elev_to_hidrology(
  elev_sgrd = NULL,
  slope_sgrd = NULL,
  spcar_sgrd = NULL,
  out_dir,
  prefix = "",
  envir,
  ...,
  mridx = FALSE,
  sagaw = FALSE,
  sllgt = FALSE,
  spidx = FALSE
)
}
\arguments{
\item{elev_sgrd}{optional input, elevation raster data in SAGA format,
can be created with \code{elev_to_sgrd()}}

\item{slope_sgrd}{optional input, slope raster data in SAGA format,
can be created with \code{elev_to_morphometry(slope = TRUE)}}

\item{spcar_sgrd}{optional input, specific catchment area raster data in
SAGA format, can be created with
\code{elev_to_terrain_analysis(flow = TRUE, spcar = TRUE)}
Both options are needed since this is computed from the
Flow Accumulation (One Step) module}

\item{out_dir}{output directory}

\item{prefix}{character prefix for output filenames}

\item{envir}{environment to get SAGA installation,
can be set with \code{init_saga()}}

\item{...}{ignored, check help page for possible outputs}

\item{mridx}{Melton Ruggedness Index:
Simple flow accumulation related index, calculated as
difference between maximum and minimum elevation in catchment
area divided by square root of catchment area size. The
calculation is performed for each grid cell, therefore minimum
elevation is same as elevation at cell’s position. Due to the
discrete character of a single maximum elevation, flow
calculation is simply done with Deterministic 8.
Boolean,defaults to \code{FALSE}}

\item{sagaw}{SAGA Wetness Index:
Similar to the ‘Topographic Wetness Index’ (TWI),
see \code{eval_to_terrain_analysis} but based on a modified
catchment area calculation (‘Modified Catchment Area’),
which does not think of the flow as very thin film. It predicts
for cells situated in valley floors with a small vertical
distance to a channel a more realistic, higher potential
soil moisture compared to the standard TWI calculation.
Boolean, defaults to \code{FALSE}}

\item{sllgt}{Slope Length:
Measurement of the distance from the origin of overland flow
along its flow path to the location of either concentrated flow
or deposition.
Boolean, defaults to \code{FALSE}}

\item{spidx}{Stream Power Index:
Measure of the erosive power of flowing water. Calculated based
upon slope and contributing area. SPI approximates locations
where gullies might be more likely to form on the landscape.
Needs \code{slope_sgrd} and \code{spcar_sgrd} parameters.
Boolean, defaults to \code{FALSE}}
}
\description{
Calculate derivatives from the Hidrology module
}
