% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting.R
\name{elev_to_lighting}
\alias{elev_to_lighting}
\title{Calculate derivatives from the Lighting module}
\usage{
elev_to_lighting(
  elev_sgrd,
  out_dir,
  prefix = "",
  env = env,
  ...,
  negop = FALSE,
  posop = FALSE,
  shade = FALSE,
  svfct = FALSE,
  visky = FALSE
)
}
\arguments{
\item{elev_sgrd}{input, elevation raster data in SAGA format,
can be created with \code{elev_to_sgrd()}}

\item{out_dir}{output directory}

\item{prefix}{character prefix for output filenames}

\item{env}{environment to get SAGA installation,
can be set with \code{init_saga()}}

\item{...}{ignored, check help page for possible outputs}

\item{posop, negop}{Topographic Openness:
Topographic openness expresses the dominance (positive) or
enclosure (negative) of a landscape location. Openness has been
related to how wide a landscape can be viewed from any position.
It has been proven to be a meaningful input for computer aided
geomorphological mapping. Openness is an angular measure of the
relation between surface relief and horizontal distance. For
angles less than 90", it is equivalent to the internal angle of
a cone, its apex at a DEM location, constrained by neighboring
elevations within a specified radial distance. Openness
incorporates the terrain line-of-sight, or viewshed, concept
and is calculated from multiple zenith and nadir angles-here
along eight azimuths. Openness has two viewer perspectives.
Positive values, expressing openness above the surface, are
high for convex forms, whereas negative values describe this
attribute below the surface and are high for concave forms.
Openness values are mapped by gray-scale tones. The emphasis
of terrain convexity and concavity in openness maps facilitates
the interpretation of landforms on the surface of the Earth and
its seafloor, and on the planets, as well as features on any
irregular surface-such as those generated by industrial
procedures.
Boolean, defaults to \code{FALSE}}

\item{shade}{Hillshade:
shaded relief from a surface raster by considering the
illumination source angle and shadows.
Boolean, defaults to \code{FALSE}}

\item{svfct, visky}{Sky View Factor and Visible Sky:
Crucial variable widely used to quantify the characteristics
of surface structures and estimate surface radiation budget.
The SVF expresses the proportion (ratio) of radiation leaving
the sky, assumed isotropic, that is able to reach a ground
surface tilted at an arbitrary angle. Its value must vary
between the minimum of 0, when the sky is not visible at all
and to the maximum of 1, when the ground surface is horizontal
and the sky entirely visible. Visible Sky is computed within
the same algorithm.
Boolean, defaults to \code{FALSE}}
}
\description{
Calculate derivatives from the Lighting module
}
